flask_migration: merge pages.utils and page.utils_flask
