+++ inginious/frontend/pages/group.py
@@ -11,7 +11,7 @@ from flask import request, redirect
 from werkzeug.exceptions import Forbidden
 from bson.objectid import ObjectId
 
-from inginious.frontend.pages.utils_flask import INGIniousAuthPage
+from inginious.frontend.pages.utils import INGIniousAuthPage
 
 
 class GroupPage(INGIniousAuthPage):
