+++ inginious/frontend/pages/marketplace.py
@@ -12,7 +12,7 @@ from inginious.common.base import id_checker
 from inginious.common.exceptions import ImportCourseException
 from inginious.common.log import get_course_logger
 from inginious.frontend.marketplace_courses import get_all_marketplace_courses, get_marketplace_course
-from inginious.frontend.pages.utils_flask import INGIniousAuthPage
+from inginious.frontend.pages.utils import INGIniousAuthPage
 
 if sys.platform == 'win32':
     import pbs
