+++ inginious/frontend/pages/preferences/bindings.py
@@ -7,7 +7,7 @@
 from flask import request, redirect
 
 from pymongo import ReturnDocument
-from inginious.frontend.pages.utils_flask import INGIniousAuthPage
+from inginious.frontend.pages.utils import INGIniousAuthPage
 
 
 class BindingsPage(INGIniousAuthPage):
