+++ inginious/frontend/pages/utils.py
@@ -7,11 +7,13 @@
 import logging
 from typing import List, Dict
 
-import web
 import os
 from gridfs import GridFS
-from inginious.client.client import Client
+from flask import redirect, request, current_app, session, url_for
+from flask.views import MethodView
+from werkzeug.exceptions import NotFound, NotAcceptable
 
+from inginious.client.client import Client
 from inginious.common import custom_yaml
 from inginious.frontend.environment_types import get_all_env_types
 from inginious.frontend.environment_types.env_type import FrontendEnvType
@@ -27,7 +29,7 @@ from inginious.frontend.task_factory import TaskFactory
 from inginious.frontend.lti_outcome_manager import LTIOutcomeManager
 
 
-class INGIniousPage(object):
+class INGIniousPage(MethodView):
     """
     A base for all the pages of the INGInious webapp.
     Contains references to the PluginManager, the CourseFactory, and the SubmissionManager
@@ -41,7 +43,31 @@ class INGIniousPage(object):
     @property
     def app(self):
         """ Returns the web application singleton """
-        return web.ctx.app_stack[0]
+        return current_app
+
+    def _pre_check(self, sessionid):
+        # Check for cookieless redirect
+        if not sessionid and session.get("cookieless", False):
+            query_string = "?" + request.query_string.decode("utf-8") if request.query_string else ""
+            request.view_args.update(sessionid=session.get("session_id"))
+            return redirect(url_for(request.endpoint, **request.view_args) + query_string)
+
+        # Check for language
+        if "lang" in request.args and request.args["lang"] in self.app.l10n_manager.translations.keys():
+            self.user_manager.set_session_language(request.args["lang"])
+        elif "language" not in session:
+            best_lang = request.accept_languages.best_match(self.app.l10n_manager.translations.keys(), default="en")
+            self.user_manager.set_session_language(best_lang)
+
+        return ""
+
+    def get(self, sessionid, *args, **kwargs):
+        pre_check = self._pre_check(sessionid)
+        return pre_check if pre_check else self.GET(*args, **kwargs)
+
+    def post(self, sessionid, *args, **kwargs):
+        pre_check = self._pre_check(sessionid)
+        return pre_check if pre_check else self.POST(*args, **kwargs)
 
     @property
     def plugin_manager(self) -> PluginManager:
@@ -140,10 +166,10 @@ class INGIniousAuthPage(INGIniousPage):
     """
 
     def POST_AUTH(self, *args, **kwargs):  # pylint: disable=unused-argument
-        raise web.notacceptable()
+        raise NotAcceptable()
 
     def GET_AUTH(self, *args, **kwargs):  # pylint: disable=unused-argument
-        raise web.notacceptable()
+        raise NotAcceptable()
 
     def GET(self, *args, **kwargs):
         """
@@ -155,7 +181,7 @@ class INGIniousAuthPage(INGIniousPage):
                                                              self.app.privacy_page is not None and
                                                              not self.user_manager.session_tos_signed()))\
                     and not self.__class__.__name__ == "ProfilePage":
-                raise web.seeother("/preferences/profile")
+                return redirect("/preferences/profile")
 
             if not self.is_lti_page and self.user_manager.session_lti_info() is not None: #lti session
                 self.user_manager.disconnect_user()
@@ -166,10 +192,10 @@ class INGIniousAuthPage(INGIniousPage):
             return self.GET_AUTH(*args, **kwargs)
         else:
             error = ''
-            if "binderror" in web.input():
+            if "binderror" in request.args:
                 error = _("An account using this email already exists and is not bound with this service. "
                           "For security reasons, please log in via another method and bind your account in your profile.")
-            if "callbackerror" in web.input():
+            if "callbackerror" in request.args:
                 error = _("Couldn't fetch the required information from the service. Please check the provided "
                           "permissions (name, email) and contact your INGInious administrator if the error persists.")
             return self.template_helper.render("auth.html", auth_methods=self.user_manager.get_auth_methods(), error=error)
@@ -181,7 +207,7 @@ class INGIniousAuthPage(INGIniousPage):
         """
         if self.user_manager.session_logged_in():
             if not self.user_manager.session_username() and not self.__class__.__name__ == "ProfilePage":
-                raise web.seeother("/preferences/profile")
+                return redirect("/preferences/profile")
 
             if not self.is_lti_page and self.user_manager.session_lti_info() is not None:  # lti session
                 self.user_manager.disconnect_user()
@@ -189,7 +215,7 @@ class INGIniousAuthPage(INGIniousPage):
 
             return self.POST_AUTH(*args, **kwargs)
         else:
-            user_input = web.input()
+            user_input = request.form
             if "login" in user_input and "password" in user_input:
                 if self.user_manager.auth_user(user_input["login"].strip(), user_input["password"]) is not None:
                     return self.GET_AUTH(*args, **kwargs)
@@ -210,15 +236,62 @@ class INGIniousAuthPage(INGIniousPage):
 
 class SignInPage(INGIniousAuthPage):
     def GET_AUTH(self, *args, **kwargs):
-        raise web.seeother("/mycourses")
+        return redirect("/mycourses")
 
     def POST_AUTH(self, *args, **kwargs):
-        raise web.seeother("/mycourses")
+        return redirect("/mycourses")
 
     def GET(self):
         return INGIniousAuthPage.GET(self)
 
 
+class LogOutPage(INGIniousAuthPage):
+    def GET_AUTH(self, *args, **kwargs):
+        self.user_manager.disconnect_user()
+        return redirect("/courselist")
+
+    def POST_AUTH(self, *args, **kwargs):
+        self.user_manager.disconnect_user()
+        return redirect("/courselist")
+
+
+class INGIniousStaticPage(INGIniousPage):
+    cache = {}
+
+    def GET(self, pageid):
+        return self.show_page(pageid)
+
+    def POST(self, pageid):
+        return self.show_page(pageid)
+
+    def show_page(self, page):
+        static_directory = self.app.static_directory
+        language = self.user_manager.session_language()
+
+        # Check for the file
+        filename = None
+        mtime = None
+        filepaths = [os.path.join(static_directory, page + ".yaml"),
+                     os.path.join(static_directory, page + "." + language + ".yaml")]
+
+        for filepath in filepaths:
+            if os.path.exists(filepath):
+                filename = filepath
+                mtime = os.stat(filepath).st_mtime
+
+        if not filename:
+            raise NotFound(description=_("File doesn't exist."))
+
+        # Check and update cache
+        if INGIniousStaticPage.cache.get(filename, (0, None))[0] < mtime:
+            with open(filename, "r") as f:
+                INGIniousStaticPage.cache[filename] = mtime, custom_yaml.load(f)
+
+        filecontent = INGIniousStaticPage.cache[filename][1]
+        title = filecontent["title"]
+        content = ParsableText.rst(filecontent["content"], initial_header_level=2)
+
+        return self.template_helper.render("static.html", pagetitle=title, content=content)
 
 
 def generate_user_selection_box(user_manager: UserManager, render_func, current_users: List[str], course_id: str, name: str, id:str, placeholder:str=None, single=False):
