+++ inginious/frontend/plugins/auth/saml2_auth.py
@@ -16,7 +16,7 @@ from werkzeug.exceptions import abort, InternalServerError
 from onelogin.saml2.auth import OneLogin_Saml2_Auth
 from onelogin.saml2.utils import OneLogin_Saml2_Utils
 
-from inginious.frontend.pages.utils_flask import INGIniousPage
+from inginious.frontend.pages.utils import INGIniousPage
 from inginious.frontend.user_manager import AuthMethod
 
 settings = None
